Web for clients

+++ src/main/java/net/proselyte/trpo/service/ClientService.java
@@ -4,9 +4,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.proselyte.trpo.dto.ClientDTO;
 import net.proselyte.trpo.entity.Client;
+import net.proselyte.trpo.entity.Reservation;
+import net.proselyte.trpo.exceptions.IsNotAvailableException;
 import net.proselyte.trpo.exceptions.NoSuchException;
 import net.proselyte.trpo.mapper.ClientMapper;
 import net.proselyte.trpo.repository.ClientRepository;
+import net.proselyte.trpo.repository.ReservationRepository;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,6 +24,8 @@ import java.util.stream.Collectors;
 public class ClientService {
     private final ClientRepository clientRepository;
     private final ClientMapper clientMapper;
+    private final PasswordEncoder passwordEncoder;
+    private final ReservationRepository reservationRepository;
 
     @Transactional
     public List<ClientDTO> getAllClients() {
@@ -38,13 +44,34 @@ public class ClientService {
         return clientMapper.toDto(clientOptional.get());
     }
 
+
+    @Transactional
+    public Client registerClient(Client client) {
+        // Проверка, существует ли уже пользователь с таким email
+        if (clientRepository.findByEmail(client.getEmail()).isPresent()) {
+            throw new NoSuchException("User with email " + client.getEmail() + " already exists.");
+        }
+
+        // Хеширование пароля
+        client.setPassword(passwordEncoder.encode(client.getPassword()));
+
+        // Сохранение пользователя
+        return clientRepository.save(client);
+    }
+
     @Transactional
     public void deleteClient(Integer clientId) {
         log.info("Delete client");
         if(clientRepository.findById(clientId).isEmpty()){
             throw new NoSuchException("There is no client with ID = "+ clientId + " in Database");
         }
-        clientRepository.deleteById(clientId);
+        Reservation reservation = reservationRepository.findByClientId(clientId);
+        if(reservation == null){
+            clientRepository.deleteById(clientId);
+        }else{
+            throw new IsNotAvailableException("Detected reservation, client can`t be deleted");
+        }
+
     }
 
     @Transactional
@@ -58,6 +85,8 @@ public class ClientService {
             existingClient.setEmail(updatedClient.getEmail());
             existingClient.setPassword(updatedClient.getPassword());
             existingClient.setLastname(updatedClient.getLastname());
+            existingClient.setStatus(updatedClient.getStatus());
+            existingClient.setRole(updatedClient.getRole());
 
             return clientMapper.toDto(clientRepository.save(existingClient));
         }
